# docs/swagger.yaml
definitions:
  models.CommonRes:
    properties:
      error_code:
        type: integer
      message:
        type: string
    type: object
  models.LiveCancelParameters:
    properties:
      channel_key:
        type: string
      content_provider_key:
        type: string
      stream_keys:
        items:
          type: string
        type: array
    required:
    - channel_key
    - content_provider_key
    type: object
  models.LiveParameters:
    properties:
      channel_key:
        type: string
      content_provider_key:
        type: string
      stream_keys:
        items:
          type: string
        type: array
    required:
    - channel_key
    - content_provider_key
    - stream_keys
    type: object
  models.LivePrepareSuccessRes:
    properties:
      channel_key:
        type: string
      error_code:
        type: string
      message:
        type: string
      stream_keys:
        items:
          type: string
        type: array
    type: object
...
  models.NginxBroadCastCloseReq:
    properties:
      broadcast_key:
        type: string
      channel_key:
        type: string
    required:
    - broadcast_key
    - channel_key
    type: object
  models.NginxBroadCastOpenRes:
    properties:
      audio_profile:
        type: object
      broadcast_key:
        type: string
      channel_key:
        type: string
      client:
        type: string
      error_code:
        type: string
      listen_port:
        type: integer
      servers:
        items:
          type: object
        type: array
      stream_plugin_id:
        type: integer
      video_profiles:
        items:
          type: object
        type: array
    type: object
  models.NginxParameters:
    properties:
      broadcast_key:
        type: string
      channel_key:
        type: string
      client:
        type: string
      host:
        type: string
      record:
        default: false
        type: boolean
      rtmp:
        type: string
      server:
        type: string
      stream_key:
        type: string
    required:
    - channel_key
    - client
    - host
    - rtmp
    - server
    - stream_key
    type: object
...
info:
  contact:
    email: deuksoo.mun@catenoid.net
    name: API Support
  description: Webrtc Controller Api
  title: WebRtc Controller Swagger API
  version: "1.0"
paths:
  /channel:
    delete:
      consumes:
      - application/json
      description: Send to cancel signal to nginx agent
      parameters:
      - description: Send to cancel signal to nginx agent
        in: body
        name: BroadCastPrepareInfo
        required: true
        schema:
          $ref: '#/definitions/models.LiveCancelParameters'
      produces:
      - application/json
      responses:
        "204":
          description: The ChannelKey cancel on prepared broadcast
          schema:
            type: "null"
        "400":
          description: Invalid Parameters
          schema:
            $ref: '#/definitions/models.CommonRes'
        "401":
          description: Validation Authorization
          schema:
            $ref: '#/definitions/models.CommonRes'
        "408":
          description: Nginx Agent Request is timeout
          schema:
            $ref: '#/definitions/models.CommonRes'
        "412":
          description: All Nginx Agent Request isn't success
          schema:
            $ref: '#/definitions/models.CommonRes'
        "417":
          description: Webrtc Controller Server Exception
          schema:
            $ref: '#/definitions/models.CommonRes'
      summary: Cancel Prepared BroadCast Channel
      tags:
      - Broadcast
    post:
      consumes:
      - application/json
      description: Prepare to Broadcast and Send to ready signal to nginx agent
      parameters:
      - description: Prepare to Broadcast and Send to ready signal to nginx agent
        in: body
        name: BroadCastPrepareInfo
        required: true
        schema:
          $ref: '#/definitions/models.LiveParameters'
      produces:
      - application/json
      responses:
        "200":
          description: The ChannelKey Prepared Broadcast
          schema:
            $ref: '#/definitions/models.LivePrepareSuccessRes'
        "400":
          description: Invalid Parameters
          schema:
            $ref: '#/definitions/models.CommonRes'
        "401":
          description: Validation Authorization
          schema:
            $ref: '#/definitions/models.CommonRes'
        "408":
          description: Nginx Agent Request is timeout
          schema:
            $ref: '#/definitions/models.CommonRes'
        "412":
          description: All Nginx Agent Request isn't success
          schema:
            $ref: '#/definitions/models.CommonRes'
        "417":
          description: Webrtc Controller Server Exception
          schema:
            $ref: '#/definitions/models.CommonRes'
      summary: Prepared BroadCast Channel
      tags:
      - Broadcast
...
  /publish:
    post:
      consumes:
      - application/json
      description: BroadCast is start. All Agent and live Stream Api send to broadcast
        open signal
      parameters:
      - description: BroadCast is start. All Agent and live Stream Api send to broadcast
          open signal
        in: body
        name: BroadCastStartInfo
        required: true
        schema:
          $ref: '#/definitions/models.NginxParameters'
      produces:
      - application/json
      responses:
        "200":
          description: The ChannelKey Start Broadcast
          schema:
            allOf:
            - $ref: '#/definitions/models.NginxBroadCastOpenRes'
            - properties:
                audio_profile:
                  allOf:
                  - type: object
                  - properties:
                      audio_bitrate:
                        type: integer
                      audio_codec:
                        type: string
                      audio_port:
                        type: integer
                      audio_sample_rate:
                        type: integer
                    type: object
                servers:
                  items:
                    allOf:
                    - type: object
                    - properties:
                        ip:
                          type: string
                      type: object
                  type: array
                video_profiles:
                  items:
                    allOf:
                    - type: object
                    - properties:
                        video_bitrate:
                          type: integer
                        video_codec:
                          type: string
                        video_framerate:
                          type: integer
                        video_height:
                          type: integer
                        video_port:
                          type: integer
                        video_width:
                          type: integer
                      type: object
                  type: array
              type: object
        "400":
          description: Invalid Parameters
          schema:
            $ref: '#/definitions/models.CommonRes'
        "401":
          description: Validation Authorization
          schema:
            $ref: '#/definitions/models.CommonRes'
        "412":
          description: The response requested to another agent has failed
          schema:
            $ref: '#/definitions/models.CommonRes'
        "417":
          description: Webrtc Controller Server Exception
          schema:
            $ref: '#/definitions/models.CommonRes'
      summary: Register BroadCast
      tags:
      - Broadcast
  /unPublish:
    post:
      consumes:
      - application/json
      description: BroadCast is close. All Agent and live Stream Api send to broadcast
        close signal
      parameters:
      - description: BroadCast is close. All Agent and live Stream Api send to broadcast
          close signal
        in: body
        name: BroadCastCloseInfo
        required: true
        schema:
          $ref: '#/definitions/models.NginxBroadCastCloseReq'
      produces:
      - application/json
      responses:
        "200":
          description: The Broadcast success to closed
          schema:
            $ref: '#/definitions/models.CommonRes'
        "206":
          description: The Broadcast is success to closed partially
          schema:
            $ref: '#/definitions/models.CommonRes'
        "400":
          description: Invalid Parameters
          schema:
            $ref: '#/definitions/models.CommonRes'
        "401":
          description: Validation Authorization
          schema:
            $ref: '#/definitions/models.CommonRes'
        "408":
          description: Janus Agent Request is timeout
          schema:
            $ref: '#/definitions/models.CommonRes'
        "417":
          description: Webrtc BroadcastKey wasn't deleted
          schema:
            $ref: '#/definitions/models.CommonRes'
      summary: Close BroadCast
      tags:
      - Broadcast
swagger: "2.0"